More improvements to setup_linux_on_wsl

+++ bin/setup_linux_on_wsl.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 set -x
 
+export CF_DEPLOYMENT_VERSION=v4.2.0
 export SCRIPT="$( basename "${BASH_SOURCE[0]}" )"
 export SCRIPTPATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
@@ -58,6 +59,9 @@ function cloneRepo() {
 
 function installBosh() {
     $REPOS_DIR/solace-messaging-cf-dev/bin/bosh_lite_vm.sh -c
+    if [ ! -e /usr/local/bin/bosh ]; then
+        sudo cp $WORKSPACE/bucc/bin/bosh /usr/local/bin
+    fi
 }
 
 function deployCf() {
@@ -67,20 +71,22 @@ function deployCf() {
 
 function installPrograms() {
 
-    # Install the cf cli tool.
-    curl -L "https://packages.cloudfoundry.org/stable?release=linux64-binary&source=github" | tar -zx
-    sudo mv cf /usr/local/bin
+    if [ ! -e /usr/local/bin/cf ]; then
+        # Install the cf cli tool.
+        curl -L "https://packages.cloudfoundry.org/stable?release=linux64-binary&source=github" | tar -zx
+        sudo mv cf /usr/local/bin
 
-    sudo apt-get update
+        sudo apt-get update
 
-    sudo apt-get install -y jq build-essential zlibc zlib1g-dev ruby ruby-dev rubygems openssl libssl-dev libxslt-dev libxml2-dev libreadline-dev libyaml-dev libsqlite3-dev sqlite3
-    sudo gem install bundler
+        sudo apt-get install -y jq build-essential zlibc zlib1g-dev ruby ruby-dev rubygems openssl libssl-dev libxslt-dev libxml2-dev libreadline-dev libyaml-dev libsqlite3-dev sqlite3
+        sudo gem install bundler
+    fi
 }
 
 function getSettingsEnv() {
-		echo "export SOLACE_MESSAGING_CF_DEV=$REPOS_DIR/solace-messaging-cf-dev"
-		echo "export WORKSPACE=$HOME/workspace"
-		echo "export PATH=\$PATH:$WORKSPACE/bucc/bin"
+    echo "export SOLACE_MESSAGING_CF_DEV=$REPOS_DIR/solace-messaging-cf-dev"
+    echo "export WORKSPACE=$HOME/workspace"
+    echo "export PATH=\$PATH:$WORKSPACE/bucc/bin"
 }
 
 function createSettingsFile() {
@@ -95,7 +101,6 @@ function alterProfile() {
     NUM_LINES=$( grep -c "source $SETTINGS_FILE" ~/.profile )
 
     if [ "$NUM_LINES" -eq 0 ]; then
-        echo out there
         read -p "Would you like  your .profile modified to automatically set up the CF environment when you next log in? (yN): "
 
         if [[ $REPLY =~ ^[Yy] ]]; then
@@ -107,17 +112,16 @@ function alterProfile() {
 
 function setupLinuxOnWsl() {
 
-cd
-setupLinks
-installPrograms
-set -e
-cloneRepo
-installBosh
-deployCf
-createSettingsFile
-set +e
-alterProfile
-
+    cd
+    setupLinks
+    installPrograms
+    set -e
+    cloneRepo
+    installBosh
+    deployCf
+    createSettingsFile
+    set +e
+    alterProfile
 }
